// renderer/utils/types/responses/receipts/create.ts
export type ReceiptCreateResponse = {
	id?: number;
	receipt_no?: string;
	student_id: number;
	student_name: string;
	form_id: number;
	form_name: string;
	date: Date;
	payment_year: number;
	jan: number;
	feb: number;
	mar: number;
	apr: number;
	may: number;
	jun: number;
	jul: number;
	aug: number;
	sep: number;
	oct: number;
	nov: number;
	dec: number;
	class_fees_per_month: number;
	package_discount_per_month: number;
	total_class_fees: number;
	total_package_discount: number;
	reg_fees: number;
	incentive: number;
	voucher_id: string;
	voucher_discount: number;
	voucher_desc: string;
	tax_amount: number;
	tax_desc: string;
	subtotal: number;
	total: number;
	remarks?: string;
	status?: string;
	receipt_class: {
		class_id: number;
		class_name: string;
		teacher_id: number;
		teacher_name: string;
		fees: number;
		package_discount: number;
		is_package: boolean;
	}[];
};